refactor: share DOM stubs across tests
