+++ test/match.test.js
@@ -1,28 +1,9 @@
 import test from 'node:test';
 import assert from 'node:assert/strict';
 import { mulberry32, state } from '../src/state.js';
+import { stubMatchDom } from './helpers/dom-stubs.js';
 
-// minimal DOM stubs for modules that expect a browser environment
-global.window = { devicePixelRatio: 1 };
-const canvasEl = {
-  getContext: () => ({ setTransform(){}, measureText(){ return { width: 0 }; }, clearRect(){} }),
-  width: 980,
-  height: 560,
-  style: {},
-  parentElement: { offsetWidth: 980 },
-  getBoundingClientRect: () => ({ left: 0, top: 0 })
-};
-const statusEl = { textContent: '' };
-const srEl = { innerHTML: '' };
-
-global.document = {
-  getElementById: (id) => {
-    if(id === 'canvas') return canvasEl;
-    if(id === 'status') return statusEl;
-    if(id === 'sr-game-state') return srEl;
-    return canvasEl;
-  }
-};
+const { statusEl } = stubMatchDom();
 
 const { startMatchRound } = await import('../src/match.js');
